Move ops to top-level in subcrate.
