--- src/daft-execution/src/compute/mod.rs
@@ -1,1 +1,0 @@
-pub mod ops;
