+++ src/daft-execution/src/ops/scan.rs
